Prevent fruit spawning on top of the snake.
This closes #5. The solution is a but gnarly, and the game will definitely crash if the player manages a perfect game and occupies the whole board... but that's probably gonna happen anyway. Maybe a compromise is to throw a particular exception and rescue that so we can save the highscore??... Imma open an issue.